# feat: Update default table storage query (INT1-106)